requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
debtcollector==1.20.0
defusedxml==0.5.0
Django==2.1.7
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-csv==2.1.0
djangorestframework-gis==0.14
djangorestframework-xml==1.4.0
drf-amsterdam==0.1.8
drf-extensions==0.4.0
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
factory-boy==2.11.1
Faker==1.0.1
graypy==0.3.1
idna==2.8
ipaddress==1.0.22
iso8601==0.1.12
itypes==1.1.0
Jinja2==2.10
keystoneauth1==3.11.2
MarkupSafe==1.1.0
monotonic==1.5
msgpack==0.6.0
netaddr==0.7.19
netifaces==0.10.9
openapi-codec==1.3.2
os-service-types==1.4.0
oslo.config==6.7.0
oslo.i18n==3.23.0
oslo.serialization==2.28.1
oslo.utils==3.39.0
pbr==5.1.1
psycopg2-binary==2.7.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
python-keystoneclient==3.18.0
python-swiftclient==3.6.0
pytz==2018.9
PyYAML==4.2b4
requests==2.21.0
rfc3986==1.2.0
sentry-sdk==0.6.6
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
stevedore==1.30.0
text-unidecode==1.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
